third_party/blink/web_tests/fast/media/mq-display-mode-expected.txt

Test the (display-mode: ...) media features.

------------- with 'browser' display mode -------------
Query "(display-mode)": true
Query "(display-mode: browser)": true
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": false
Query "(display-mode: window-controls-overlay)": false
Query "(display-mode: borderless)": false
Query "(display-mode: fine)": false
Query "(display-mode: 1)": false
Query "(display-mode: 3%)": false
------------- with 'minimal-ui' display mode -------------
Query "(display-mode)": true
Query "(display-mode: browser)": false
Query "(display-mode: minimal-ui)": true
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": false
Query "(display-mode: window-controls-overlay)": false
Query "(display-mode: borderless)": false
Query "(display-mode: fine)": false
Query "(display-mode: 1)": false
Query "(display-mode: 3%)": false
------------- with 'standalone' display mode -------------
Query "(display-mode)": true
Query "(display-mode: browser)": false
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": true
Query "(display-mode: fullscreen)": false
Query "(display-mode: window-controls-overlay)": false
Query "(display-mode: borderless)": false
Query "(display-mode: fine)": false
Query "(display-mode: 1)": false
Query "(display-mode: 3%)": false
------------- with 'fullscreen' display mode -------------
Query "(display-mode)": true
Query "(display-mode: browser)": false
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": true
Query "(display-mode: window-controls-overlay)": false
Query "(display-mode: borderless)": false
Query "(display-mode: fine)": false
Query "(display-mode: 1)": false
Query "(display-mode: 3%)": false
------------- with 'window-controls-overlay' display mode -------------
Query "(display-mode)": true
Query "(display-mode: browser)": false
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": false
Query "(display-mode: window-controls-overlay)": true
Query "(display-mode: borderless)": false
Query "(display-mode: fine)": false
Query "(display-mode: 1)": false
Query "(display-mode: 3%)": false
------------- with 'borderless' display mode -------------
Query "(display-mode)": true
Query "(display-mode: browser)": false
Query "(display-mode: minimal-ui)": false
Query "(display-mode: standalone)": false
Query "(display-mode: fullscreen)": false
Query "(display-mode: window-controls-overlay)": false
Query "(display-mode: borderless)": true
Query "(display-mode: fine)": false
Query "(display-mode: 1)": false
Query "(display-mode: 3%)": false
